+++ Source/MonolithCore/Public/MonolithCoreModule.h
@@ -2,7 +2,7 @@
 
 #include "Modules/ModuleManager.h"
 
-#define MONOLITH_VERSION TEXT("0.20.1")
+#define MONOLITH_VERSION TEXT("0.20.2")
 
 class FMonolithHttpServer;